# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Campus Lost & Found System (CLAFS) API",
    "version": "1.0.0",
    "description": "RESTful API for CLAFS — a digital lost-and-found platform for university campuses. Exposes endpoints for user management, item report CRUD, and the full claim workflow (submit → review → approve/reject).\n\nAll responses follow a consistent envelope:\n- **success**: `{ status: 'success', data: <payload> }`\n- **list**: `{ status: 'success', count: N, data: [<payload>] }`\n- **error**: `{ status: 'error', error: '<ErrorType>', message: '<detail>' }`",
    "contact": {
      "name": "Mlungisi Mbuyazi",
      "email": "admin@cput.ac.za"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/users/register": {
      "post": {
        "tags": ["Users"],
        "summary": "Register a new user",
        "description": "Creates a new user account. Email must be a valid university address (e.g. `@cput.ac.za`, `@edu`, `@ac.uk`). Email is immutable after registration.",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" },
              "example": {
                "name": "Thabo Nkosi",
                "email": "thabo@cput.ac.za",
                "password": "SecurePass1",
                "role": "STUDENT"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "409": { "$ref": "#/components/responses/ConflictError" }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login",
        "description": "Authenticates a user by email and password. Accounts are locked for 15 minutes after 3 consecutive failed attempts.",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/AuthError" }
        }
      }
    },
...
    "/api/users/role/{role}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get users by role",
        "operationId": "getUsersByRole",
        "parameters": [
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "enum": ["STUDENT", "LECTURER", "ADMIN"] }
          }
        ],
        "responses": {
          "200": { "description": "Filtered user list" },
          "400": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/api/users/{userId}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a user by ID",
        "operationId": "getUserById",
        "parameters": [{ "$ref": "#/components/parameters/userId" }],
        "responses": {
          "200": { "description": "User found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user profile",
        "description": "Updates name and/or role. Email cannot be changed after registration.",
        "operationId": "updateUser",
        "parameters": [{ "$ref": "#/components/parameters/userId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "minLength": 2 },
                  "role": { "type": "string", "enum": ["STUDENT", "LECTURER", "ADMIN"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete a user",
        "operationId": "deleteUser",
        "parameters": [{ "$ref": "#/components/parameters/userId" }],
        "responses": {
          "200": { "description": "User deleted" },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      }
    },
    "/api/users/{userId}/verify": {
      "patch": {
        "tags": ["Users"],
        "summary": "Verify user email",
        "operationId": "verifyEmail",
        "parameters": [{ "$ref": "#/components/parameters/userId" }],
        "responses": {
          "200": { "description": "Email verified" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "409": { "$ref": "#/components/responses/ConflictError" }
        }
      }
    },
...
    "/api/reports/user/{userId}": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get reports by user",
        "operationId": "getReportsByUser",
        "parameters": [{ "$ref": "#/components/parameters/userId" }],
        "responses": {
          "200": { "description": "Reports for user" },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      }
    },
...
    "/api/claims/user/{claimantId}": {
      "get": {
        "tags": ["Claims"],
        "summary": "Get claims by a specific user",
        "operationId": "getClaimsByUser",
        "parameters": [
          { "name": "claimantId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Claims by user" },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "email", "password", "role"],
        "properties": {
          "name":     { "type": "string", "minLength": 2, "example": "Thabo Nkosi" },
          "email":    { "type": "string", "format": "email", "example": "thabo@cput.ac.za" },
          "password": { "type": "string", "minLength": 8, "example": "SecurePass1" },
          "role":     { "type": "string", "enum": ["STUDENT", "LECTURER", "ADMIN"] }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email":    { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "userId":              { "type": "string", "format": "uuid" },
          "name":                { "type": "string" },
          "email":               { "type": "string", "format": "email" },
          "role":                { "type": "string", "enum": ["STUDENT", "LECTURER", "ADMIN"] },
          "isVerified":          { "type": "boolean" },
          "failedLoginAttempts": { "type": "integer" },
          "lockedUntil":         { "type": "string", "format": "date-time", "nullable": true },
          "createdAt":           { "type": "string", "format": "date-time" }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "success" },
          "data":   { "$ref": "#/components/schemas/User" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status":  { "type": "string", "example": "error" },
          "error":   { "type": "string", "example": "ValidationError" },
          "message": { "type": "string" }
        }
      }
    },
    "parameters": {
      "userId": {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "UUID of the user"
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Invalid input data",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": "error", "error": "ValidationError", "message": "Title is required and must be at least 3 characters" }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": "error", "error": "NotFoundError", "message": "ItemReport \"abc-123\" not found" }
          }
        }
      },
      "ConflictError": {
        "description": "Conflict with current resource state",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": "error", "error": "ConflictError", "message": "You have already submitted a claim for this item" }
          }
        }
      },
...
      "AuthError": {
        "description": "Authentication failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": "error", "error": "AuthError", "message": "Invalid email or password" }
          }
        }
      }
    }
  }
}